I fucked up
This code does NOT work correctly. I made the code efficient AFTER I made it work, therefore there are too many errors to search for. I'm uploading this as a base to start over from scratch and to code it correctly the FIRST time, and check correctness as it goes. Most of the algorithms and such are still good, they are just not using static variables and such. There are also a few things I wanted to create/add to make the code look a lot cleaner that just make starting over a lot easier. Therefore, we are starting over ;-;